IB/hfi1: Unregister mmu callbacks from the correct mm
The hfi1 driver registers a mmu_notifier callback when /dev/hfi1 is opened,
and unregisters it when the device is closed.  The driver assumes that the
close will always happen from the same context as the open, which is
incorrect.  In particular, closes due to SIGKILL or OOM killer activity may
happen from a different context.  In these cases, the wrong mm is passed to
mmu_notifier_unregister(), which causes improper reference counting for the
victim mm, and eventual memory corruption.

The mmu_rb_handler structure is modified to cache the mm used for successful
mmu_notifier registration, so that it can be used to unregister the
mmu_notifier.

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Jim Foraker <foraker1@llnl.gov>