IB/hfi1: Fix handling of mm in mmu_rb remove() callback
The remove() mmu_rb callback is called in hfi1_mmu_unregister() with a
NULL mm argument while holding a spinlock.  In the case of the
sdma_rb_remove() callback, this causes it to pass current->mm to
hfi1_release_user_pages(), which will then attempt to acquire the
mmap_sem under a spin lock.  Additionally, current->mm is not guaranteed
to be the proper mm to operate on, which could lead to corruption of
another tasks's pinned_vm count.

By taking mmap_sem in hfi1_mmu_unregister() and explicitly passing the
correct mm to the remove() callback, the calling conventions of the
callback can be simplified, and both of the above problems rectified.

mmu_rb_remove() in user_exp_recv.c is the only other implementation of
this callback and does not reference mm, and so is not affected.

Signed-off-by: Jim Foraker <foraker1@llnl.gov>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>