IB/hfi1: Free handler object when mmu notifier registration fails
Subsequent patches need to know that handler->mm is valid
for the life of the handler object.  If registration succeeds
the mm code has properly referenced the mm and we know that
handler->mm will be valid until after our hfi1_mmu_rb_unregister
has completed.

In addition, if the mmu_notifier_register function fails there
is no reason to keep the handler object around.

Move adding the handler to our handler list after the
registration function and free handler on failure.

There is a slight chance the handler will be called between
successful registration and adding the handler to the list.
This is benign because the invalidate calls reference the
handler from the mmu_notifier object not the list.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>

---
Changes from v1:
        Remove check for handler->mm == NULL because this will no longer
	happen in unregister.