IB/hfi1: free handler object when mmu notifier registration fails
If the mmu_notifier_register function fails there is no reason
to keep the handler object around.  Move adding the handler to
after the registration function and free handler on failure.

There is a slight chance the handler will be called between
successful registration and adding the handler to the list.
This will be benign because the invalidate calls safely check
for missing handler objects and just return.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>